requirements.txt
notion-client==2.1.0
python-dotenv==1.0.0
pyyaml==6.0
schedule==1.2.1
requests==2.31.0
